requirements.txt
numpy
pandas
openpyxl
plotly-express
nbformat==4.2.0
ipykernel
matplotlib
